refactor: move resources functions to resources.ts instead of dataset.ts
